+++ ports/ports_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 The Jaeger Authors.
+// Copyright (c) 2020 The Jaeger Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,3 +13,13 @@
 // limitations under the License.
 
 package ports
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestPortToHostPort(t *testing.T) {
+	assert.Equal(t, ":42", PortToHostPort(42))
+}